requirements.txt
# packages in environment at /Users/yevhen/anaconda3/envs/texta_app:
#
# Name                    Version                   Build  Channel
ca-certificates           2019.1.23                     0  
certifi                   2019.3.9                 py37_0  
django                    2.1.7                    py37_0  
djangorestframework       3.9.2                      py_0    conda-forge
libcxx                    4.0.1                hcfea43d_1  
libcxxabi                 4.0.1                hcfea43d_1  
libedit                   3.1.20181209         hb402a30_0  
libffi                    3.2.1                h475c297_4  
ncurses                   6.1                  h0a44026_1  
openssl                   1.1.1b               h1de35cc_1  
pip                       19.0.3                   py37_0  
python                    3.7.3                h359304d_0  
pytz                      2018.9                   py37_0  
readline                  7.0                  h1de35cc_5  
setuptools                40.8.0                   py37_0  
sqlite                    3.27.2               ha441bb4_0  
tk                        8.6.8                ha441bb4_0  
wheel                     0.33.1                   py37_0  
xz                        5.2.4                h1de35cc_4  
zlib                      1.2.11               h1de35cc_3  
